+++ packages/app-builder/public/locales/en/scenarios.json
@@ -174,5 +174,6 @@
   "edit_date.duration_unit_minutes": "minutes",
   "edit_date.duration_unit_hours": "hours",
   "edit_date.duration_unit_days": "days",
-  "enum_options": "Options:"
+  "enum_options": "Options:",
+  "nest": "Nest"
 }
