+++ packages/app-builder/src/components/Scenario/AstBuilder/RootAstBuilderNode/RootAnd.tsx
@@ -7,10 +7,12 @@ import {
 import {
   type AstBuilder,
   type EditorNodeViewModel,
+  findArgumentIndexErrorsFromParent,
   hasArgumentIndexErrorsFromParent,
 } from '@app-builder/services/editor/ast-editor';
 import {
   adaptEvaluationErrorViewModels,
+  useGetNodeEvaluationErrorMessage,
   useGetOrAndNodeEvaluationErrorMessage,
 } from '@app-builder/services/validation';
 import clsx from 'clsx';
@@ -18,6 +20,7 @@ import { Fragment } from 'react';
 
 import { ScenarioValidationError } from '../../ScenarioValidationError';
 import { AstBuilderNode } from '../AstBuilderNode/AstBuilderNode';
+import { computeLineErrors } from '../AstBuilderNode/TwoOperandsLine/TwoOperandsLine';
 import { RemoveButton } from '../RemoveButton';
 import { AddLogicalOperatorButton } from './AddLogicalOperatorButton';
 
@@ -59,6 +62,7 @@ export function RootAnd({
   viewOnly?: boolean;
 }) {
   const getEvaluationErrorMessage = useGetOrAndNodeEvaluationErrorMessage();
+  const getNodeEvaluationErrorMessage = useGetNodeEvaluationErrorMessage();
   const { nodeErrors: andNodeErrors } = separateChildrenErrors(
     rootAndViewModel.errors
   );
@@ -97,6 +101,11 @@ export function RootAnd({
           const isLastCondition =
             childIndex === rootAndViewModel.children.length - 1;
 
+          const errorMessages = adaptEvaluationErrorViewModels([
+            ...computeLineErrors(child),
+            ...findArgumentIndexErrorsFromParent(child),
+          ]).map((error) => getNodeEvaluationErrorMessage(error));
+
           return (
             <Fragment key={`condition_${child.nodeId}`}>
               {/* Row 1 */}
@@ -125,23 +134,30 @@ export function RootAnd({
                 )}
               />
 
-              <div className="px-2">
+              <div className="grid grid-cols-[1fr_30px] gap-2 pl-2">
                 <AstBuilderNode
                   builder={builder}
                   editorNodeViewModel={child}
                   viewOnly={viewOnly}
                   root
                 />
-              </div>
-              {!viewOnly && (
-                <div className="flex h-10 flex-col items-center justify-center">
-                  <RemoveButton
-                    onClick={() => {
-                      builder.remove(child.nodeId);
-                    }}
-                  />
+                {!viewOnly && (
+                  <div className="flex h-10 flex-col items-center justify-center">
+                    <RemoveButton
+                      onClick={() => {
+                        builder.remove(child.nodeId);
+                      }}
+                    />
+                  </div>
+                )}
+                <div className="flex flex-row flex-wrap gap-2">
+                  {errorMessages.map((error) => (
+                    <ScenarioValidationError key={error}>
+                      {error}
+                    </ScenarioValidationError>
+                  ))}
                 </div>
-              )}
+              </div>
             </Fragment>
           );
         })}
