+++ packages/ui-design-system/src/Switch/Switch.tsx
@@ -4,7 +4,7 @@ import { forwardRef } from 'react';
 
 export const Switch = forwardRef<
   HTMLButtonElement,
-  Omit<SwitchProps, 'classValue'> & { className?: string }
+  SwitchProps & { className?: string }
 >(({ className, ...props }, ref) => {
   return (
     <Root
